[gCore] (1) rearrange headers & declarations. (2) now u can only #include "StdInterface.h" to use the Standard gCore Interface without other headers.
